CheckboxButtongGroup now accepts icons and checkedIcons props from children. Also, abstract filter functionality to another Filter component.
