+++ pharmd-app/src/components/Basic/Checkbox Controls/CheckboxButtonGroup.js
@@ -24,30 +24,6 @@ import CheckboxButton from "./CheckboxButton";
 
 const CheckboxButtonGroup = props => {
   const { onChange, label, showLabel, error, children, color, className, checkboxClassName } = props;
-  const [value, setValue] = React.useState({
-    values: []
-  });
-
-  // If the checkbox was checked and the value is not in our array of filter values then add the value to the array
-  // If the checkbox was unchecked and the value is in our array of filter values then remove the value from the array
-  // Finally, update the filter values and call the onChange function with the new list
-  const handleChange = event => {
-    const eventVal = event.target.value;
-    let newVal = value.values;
-    const index = newVal.indexOf(eventVal);
-    if (event.target.checked) {
-      if (index < 0) {
-        newVal.push(eventVal);
-        console.log("Status Value added: ", eventVal);
-      }
-    } else if (index >= 0) {
-      newVal.splice(index, 1);
-      console.log("Status Value removed: ", eventVal, " at index ", index);
-    }
-    setValue({ values: newVal });
-    console.log("Status Value: ", newVal);
-    onChange(event, newVal);
-  };
 
   return (
     <FormControl component="fieldset" error={error} className={className}>
@@ -55,15 +31,15 @@ const CheckboxButtonGroup = props => {
       <FormGroup aria-label={label}>
         {children &&
           children.map((child, index) => {
-            console.log("Child: ",{...child});
             return cloneElement(child, {
               key: index,
               control: <CheckboxButton
-                onChange={handleChange}
+                onChange={onChange}
                 color={color}
                 className={checkboxClassName}
                 // icon={undefined}
                 icon={ child.props.icon ? child.props.icon : undefined }
+                checkedIcon={ child.props.checkedIcon ? child.props.checkedIcon : undefined }
               />
             })}
           )}
